+++ scripts/reading/render.js
@@ -258,8 +258,9 @@ async function render(index, _scale = false, magnifyingGlass = false)
 			_scale = _scale * window.devicePixelRatio// * (_scale != 1 ? 1.5 : 1); // 1.5 more scale is applied to avoid blurry text due to transform if scale is not 1
 
 			let ocImg = contentRight.querySelector(magnifyingGlass ? '.reading-lens .r-img-i'+index+' oc-img' : '.r-img-i'+index+' oc-img');
-			let originalCanvas = ocImg.querySelector('canvas');
+			if(!ocImg) return;
 
+			let originalCanvas = ocImg.querySelector('canvas');
 			if(!originalCanvas) return;
 
 			let canvas = originalCanvas.cloneNode(true);