+++ scripts/settings.js
@@ -202,7 +202,7 @@ function removeMasterFolder(key)
 		storage.set('masterFolders', masterFolders);
 
 		updateMasterFolders();
-		dom.loadIndexContentLeft(true, true);
+		dom.loadIndexContentLeft(true);
 	}
 }
 
@@ -225,7 +225,7 @@ function addMasterFolder()
 				storage.set('masterFolders', masterFolders);
 
 				updateMasterFolders();
-				dom.loadIndexContentLeft(true, true);
+				dom.loadIndexContentLeft(true);
 			}
 		}